Rename and restructure SharedTest files
