+++ samples/kaspresso-sample/src/main/kotlin/com/kaspersky/kaspressample/sharedtest/SharedTestActivity.kt
@@ -1,14 +1,14 @@
-package com.kaspersky.kaspressample.robolectric
+package com.kaspersky.kaspressample.sharedtest
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.kaspersky.kaspressample.R
 
-class RobolectricActivity : AppCompatActivity() {
+class SharedTestActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_robolectric)
+        setContentView(R.layout.activity_sharedtest)
     }
 }
